chore: add back goshims (#123)
An attempt was made to remove goshims and generate the shims with
counterfeiter. This worked in this repo, but consumers of this repo had
issues because the shims generated by counterfeiter differ slightly.

One example:
```
./main.go:333:3: cannot use &osshim.OsShim{} (value of type *"code.cloudfoundry.org/goshims/osshim".OsShim) as "code.cloudfoundry.org/existingvolumebroker/fakes/osshim".Os value in argument to existingvolumebroker.New: *"code.cloudfoundry.org/goshims/osshim".OsShim does not implement "code.cloudfoundry.org/existingvolumebroker/fakes/osshim".Os (wrong type for method Create)
                have Create(string) ("code.cloudfoundry.org/goshims/osshim".File, error)
                want Create(string) (*os.File, error)

```

Because goshims shim the os.File type, but counterfeiter does not, the
signatures no longer match.